With updates to the previous PR. This contains additional code to add a welcome message showing server date to the portfolio
