+++ portfolio/bin/src/main/appengine/app.yaml
@@ -1,0 +1,1 @@
+runtime: java11
