+++ portfolio/bin/src/main/webapp/index.html
@@ -8,11 +8,10 @@
     <!--Link to Bootstrap CDN - Using bootstap to also optimize for smaller screens-->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <link rel="stylesheet" href="style.css"/>
-    <script src="script.js"></script>
     <!--Link to font awesome icons-->
     <link rel="stylesheet" href="//use.fontawesome.com/releases/v5.0.7/css/all.css">
   </head>
-  <body onLoad="showServerTime()">
+  <body>
     <!-- Home section -->
     <section id="home">
       <!-- Main Navigation Bar-->
@@ -33,11 +32,6 @@
       </nav>
     
       <!--ABOUT SECTION-->
-
-      <div id = "welcomeSection" class="col-md-8 col-md-3 col-sm-12">
-         
-        </div>
-              
     <div class="text-center About-container">
         <div class="About-container-inner">
         <div class="window">
@@ -59,7 +53,7 @@
           </div>
          </div>
           </div>
-        </div>    
+        </div>
     </div>
     </section> 
   <!--This div marks separation between diffrent sections of my portfolio website-->
